[REFACTOR] #13 - 클린 아키텍처에 맞게 패키지 분리로 인한 파일 이동
