+++ app/src/main/java/com/example/banchan/presentation/adapter/menu/MenuViewHolder.kt
@@ -1,7 +1,7 @@
-package com.example.banchan.ui.adapter
+package com.example.banchan.presentation.adapter.menu
 
 import androidx.recyclerview.widget.RecyclerView
-import com.example.banchan.data.response.ItemModel
+import com.example.banchan.domain.model.ItemModel
 import com.example.banchan.databinding.ItemMenuBinding
 
 class MenuViewHolder(private val binding: ItemMenuBinding) :
